Finish string problems from the CtCI book
I have added the last string problem from the book Cracking the Coding
Interview. It is tested if one string is the rotation of the other one
with the constraint to make only one invocation to a isSubstring method.

Also I have added an overview of number of problems and number of
solutions to the README file.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>