Add list arithmetic algorithm
List arithmetic with respect to numbers represented as linked lists. Two
numbers can be added in this fashion when applying the paper and pencil
algorithm used for elementary arithmetic. The digits, however, have to
be stored in reversed order.

I will soon add a follow-up which deals with the case, when the digits
are stored in the correct order. Also I have added an append method to
the LinkedList class to ease up things.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>